Refactor the default bindings to a separate file
