+++ browser/src/Config.ts
@@ -8,6 +8,8 @@ import * as path from "path"
 import * as Performance from "./Performance"
 import * as Platform from "./Platform"
 
+import { applyDefaultKeyBindings } from "./Input/KeyBindings"
+
 export interface IConfigValues {
 
     "activate": (oni: Oni.Plugin.Api) => void
@@ -268,52 +270,6 @@ export class Config extends EventEmitter {
         this._activateIfOniObjectIsAvailable()
     }
 
-    private _applyDefaultKeyBindings(oni: Oni.Plugin.Api): void {
-
-        const { editors, input } = oni
-
-        input.unbindAll()
-
-        const isVisualMode = () => editors.activeEditor.mode === "visual"
-        const isNormalMode = () => editors.activeEditor.mode === "normal"
-        const isInsertOrCommandMode = () => editors.activeEditor.mode === "insert" || editors.activeEditor.mode === "cmdline_normal"
-
-        if (Platform.isLinux() || Platform.isWindows()) {
-            if (this.getValue("editor.clipboard.enabled")) {
-                input.bind("<C-c>", "editor.clipboard.yank", isVisualMode)
-                input.bind("<C-v>", "editor.clipboard.paste", isInsertOrCommandMode)
-            }
-        } else {
-            if (this.getValue("editor.clipboard.enabled")) {
-                input.bind("<M-c>", "editor.clipboard.yank", isVisualMode)
-                input.bind("<M-v>", "editor.clipboard.paste", isInsertOrCommandMode)
-            }
-        }
-
-        input.bind("<f3>", "language.formatter.formatDocument")
-        input.bind("<f12>", "oni.editor.gotoDefinition")
-        input.bind("<S-C-P>", "commands.show", isNormalMode)
-        input.bind("<C-pageup>", "oni.process.cyclePrevious")
-        input.bind("<C-pagedown>", "oni.process.cycleNext")
-
-        // QuickOpen
-        input.bind("<C-p>", "quickOpen.show", isNormalMode)
-        input.bind("<C-/>", "quickOpen.showBufferLines", isNormalMode)
-        input.bind("<enter>", "quickOpen.openFile")
-        input.bind("<C-v>", "quickOpen.openFileVertical")
-        input.bind("<C-s>", "quickOpen.openFileHorizontal")
-
-        // Completion
-        input.bind(["<enter>", "<tab>"], "completion.complete")
-        input.bind(["<down>", "<C-n>"], "completion.next")
-        input.bind(["<up>", "<C-p>"], "completion.previous")
-
-        // Menu
-        input.bind(["<down>", "<C-n>"], "menu.next")
-        input.bind(["<up>", "<C-p>"], "menu.previous")
-        input.bind("<esc>", "menu.close")
-    }
-
     private applyConfig(): void {
         let userRuntimeConfigOrError = this.getUserRuntimeConfig()
         if (isError(userRuntimeConfigOrError)) {
@@ -329,7 +285,7 @@ export class Config extends EventEmitter {
 
     private _activateIfOniObjectIsAvailable(): void {
         if (this.Config && this.Config.activate && this._oniApi) {
-            this._applyDefaultKeyBindings(this._oniApi)
+            applyDefaultKeyBindings(this._oniApi, this)
             this.Config.activate(this._oniApi)
         }
     }
